\name{census}
\alias{Tx}
\docType{data}
\title{
    Marginal totals
}
\description{
}
\usage{data("census")}
\format{
  A data frame with 1 observations on the following 9 variables.
  \describe{
    \item{\code{a16.30}}{a numeric vector}
    \item{\code{au30}}{a numeric vector}
    \item{\code{female}}{a numeric vector}
    \item{\code{male}}{a numeric vector}
    \item{\code{unemployed}}{a numeric vector}
    \item{\code{employed}}{a numeric vector}
    \item{\code{income}}{a numeric vector}
    \item{\code{rural}}{a numeric vector}
    \item{\code{urban}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data('census')
}
\keyword{datasets}
