% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getD}
\alias{getD}
\title{Weights Distance}
\usage{
getD(design_weights, simulated_weights)
}
\arguments{
\item{simulated_weights}{simulated weights}

\item{design}{weights of the survey}
}
\value{
distance_weights distance between weights
}
\description{
Computes the distance between sample design weights and estimated new
weights.
\deqn{
     D_i = \sum_j^m |w_j - d_j|
     }
}
\examples{
getD(10, 20)
getD(30, 40)
}

