% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getDist}
\alias{getDist}
\title{Weights Distance}
\usage{
getDist(design_weights, simulated_weights)
}
\arguments{
\item{design_weights}{weights of the survey}

\item{simulated_weights}{simulated weights}
}
\value{
distance_weights distance between weights
}
\description{
Computes the distance between sample design weights and estimated new
weights.
\deqn{D_i = \sum_j^m |w_j - d_j|}
}
\examples{
x <- 1:5
y <- x^2
getDist(x, y)
}

