% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getED}
\alias{getED}
\title{Error in Distribution (ED)}
\usage{
getED(population_size, simulated_weights)
}
\arguments{
\item{population_size}{known population size}

\item{simulated_weights}{simulated weights}
}
\value{
error in distribution
}
\description{
Computes the ration between absolute sum of residuals (estimated - known)
and the actual population size.
\deqn{
     ED_i = \frac{|\sum w_j - pop_i|}{pop_i}
     }
}
\examples{
getED(10, 20)
getED(30, 40)
}

