% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getEM}
\alias{getEM}
\title{Error in Margin (EM)}
\usage{
getEM(population_size, simulated_weights)
}
\arguments{
\item{population_size}{known population size}

\item{simulated}{weights}
}
\value{
error in margins
}
\description{
Computes the ration between estimated and known number of individuals or
households.
\deqn{
     EM_i = \frac{\sum w_j - pop_i}{pop_i}
     }
}
\examples{
getEM(10, 20)
getEM(30, 40)
}

