% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getMChi}
\alias{getMChi}
\title{Mean Chi-squared distance}
\usage{
getMChi(design_weights, simulated_weights, population_size)
}
\arguments{
\item{simulated_weights}{simulated weights}
}
\value{
mean chi-squared distance
}
\description{
Computes the mean Chi-squared distance between sample design weights and
estimated new weights.
\deqn{
     \theta Chi_i = \sum_j^m \frac{\left(w_j \times d_j\right)^2 }{2d_j} \div m
     }
}
\examples{
getMChi(10, 20, 5)
getMChi(30, 40, 5)
}

