% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getPearson}
\alias{getPearson}
\title{Correlation Coefficient (Pearson Correlation)}
\usage{
getPearson(observed, simulated)
}
\arguments{
\item{observed}{observed marginal sums}

\item{simulated}{simulated marginal sums}
}
\value{
Pearson correlation
}
\description{
test for correlations between simulated and expected marginal totals. This
functions implements the \code{\link[stats]{cor}} function for the
estimation of the Pearson correlation coefficient.
\deqn{
     pearson <- cor(cbind(Tx, hTx), use="complete.obs", method="pearson")
     }
}

