% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getPSAE}
\alias{getPSAE}
\title{Percentage error (PSAE)}
\usage{
getPSAE(observed, simulated, population_size)
}
\arguments{
\item{observed}{observed marginal sums}

\item{simulated}{simulated marginal sums}

\item{population_size}{known population size}
}
\value{
percentage error
}
\description{
Divides the \code{\link{SAE}} by the known population size.
\deqn{
     PAE = \sum_i^n |Tx - \hat{t}x| \div pop_i \times 100
     }
}
\examples{
getPSAE(10, 20, 5)
getPSAE(30, 40, 5)
}

