% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getR}
\alias{getR}
\title{Coefficient of determination}
\usage{
getR(observed, simulated)
}
\arguments{
\item{observed}{observed marginal sums}

\item{simulated}{simulated marginal sums}
}
\description{
Compute the r-squared coefficient of determination between simulated and
observed marginal totals. This function implements the
\code{\link[stats]{lm}} function to estimate the r-squared coefficient.
}

