% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getSAE}
\alias{getSAE}
\title{Standardized absolute error (SAE)}
\usage{
getSAE(observed, simulated, population_size)
}
\arguments{
\item{observed}{observed marginal sums}

\item{simulated}{simulated marginal sums}

\item{population_size}{known population size}
}
\value{
standardized absolute error
}
\description{
Divides the \code{\link{getTAE}} by the know population size.
\deqn{
     SAE = \sum_i^n |Tx - \hat{t}x| \div pop_i
     }
}
\examples{
getSAE(10, 20, 5)
getSAE(30, 40, 5)
}

