% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getTAE}
\alias{getTAE}
\title{Total absolute error (TAE)}
\usage{
getTAE(observed, simulated)
}
\arguments{
\item{observed}{observed marginal sums}

\item{simulated}{simulated marginal sums}
}
\value{
total absolute error
}
\description{
Computes the total absolute error as the sum of the absolute difference
between observed marginal sums and simulated marginal sums.
\deqn{
     TAE = \sum_i^n |Tx - \hat{t}x|
     }
}
\examples{
getTAE(10, 20)
getTAE(30, 40)
}

