\name{survey}
\alias{X}
\docType{data}
\title{
    Survey
}
\description{
}
\usage{data("survey")}
\format{
  A data frame with 25 observations on the following 5 variables.
  \describe{
    \item{\code{age}}{a factor with levels \code{a16.30} \code{au30}}
    \item{\code{sex}}{a factor with levels \code{female} \code{male}}
    \item{\code{employment}}{a factor with levels \code{employed} \code{unemployed}}
    \item{\code{income}}{a numeric vector}
    \item{\code{location}}{a factor with levels \code{rural} \code{urban}}
    \item{\code{weights}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data('survey')
}
\keyword{datasets}
