\name{dx}
\alias{dx}
\docType{data}
\title{
    Design weights
}
\description{
}
\usage{data("dx")}
\format{
  The format is:
 num [1:25] 4 5 6 5 3 4 6 4 5 3 ...
}
\details{
    \tabular{l}{
   4 \cr
   5 \cr
   6 \cr
   5 \cr
   \vdots
   4 \cr
   3 \cr
    }
}
\source{
    Rahman (2009)
}
\references{
    Rahman, A. (2009) Small Area Estimation Through Spatial Microsimulation Models
    2nd International Microsimulation Association Conference, Ottawa/Canada 
}
\examples{
data(dx)
}
\keyword{datasets}
