% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getChi}
\alias{getChi}
\title{Total Chi-squared distance}
\usage{
getChi(design_weights, simulated_weights)
}
\value{
chi-squared distance
}
\description{
Computes the Chi-squared distance between sample design weights and
estimated new weights.
\deqn{
     Chi_i = \sum_j^m \frac{\left( w_j \times d_j\right)^2 }{2d_j}
     }
}
\examples{
getChi(10, 20)
getChi(30, 40)
}

