% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getD}
\alias{getD}
\title{Weights Distance}
\usage{
getD(design_weights, simulated_weights)
}
\arguments{
\item{design_weights}{design weights of the survey}

\item{simulated_weights}{simulated weights}
}
\value{
distance between weights
}
\description{
Computes the weight distance
\deqn{D_i = \sum_j^m |w_j - d_j|}
}
\examples{
getD(10, 20)
getD(30, 40)
}

