% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internalValidation.R
\name{getPearson}
\alias{getPearson}
\title{Correlation Coefficient (Pearson Correlation)}
\usage{
getPearson(observed, simulated)
}
\description{
Correlation Coefficient (Pearson Correlation)
}

